+++ scripts/install/common.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+# Create logs directory
+if [ -d $DATA_DIRECTORY/logs ]; then
+    rm -rf $DATA_DIRECTORY/logs
+fi
+mkdir $DATA_DIRECTORY/logs