+++ scripts/install/php.sh
@@ -2,16 +2,6 @@
 
 echo "Configure PHP..."
 
-# PHP Configuration
-PHP_TIMEZONE=${PHP_TIMEZONE:-Europe/Paris}
-
-# Checking PHP Timezone
-if [[ ! "${PHP_TIMEZONE}" =~ ^[A-Z]{1}[a-z]+/[A-Z]{1}[a-z]+$ ]]; then
-    echo "ERROR: "
-    echo "  Variable PHP_TIMEZONE isn't valid ! (Format accepted : [A-Z]{1}[a-z]+/[A-Z]{1}[a-z]+)"
-    exit 1
-fi
-
 # Config PHP Timezone
 sed -i "s|;date.timezone =.*|date.timezone = ${PHP_TIMEZONE}|g" /etc/php5/fpm/php.ini
 sed -i "s|;date.timezone =.*|date.timezone = ${PHP_TIMEZONE}|g" /etc/php5/cli/php.ini