+++ scripts/install/toran.sh
@@ -2,62 +2,6 @@
 
 echo "Configure Toran Proxy..."
 
-# Toran Proxy Configuration
-TORAN_HOST=${TORAN_HOST:-localhost}
-TORAN_HTTP_PORT=${TORAN_HTTP_PORT:-80}
-TORAN_HTTPS=${TORAN_HTTPS:-false}
-TORAN_HTTPS_PORT=${TORAN_HTTPS_PORT:-443}
-TORAN_REVERSE=${TORAN_REVERSE:-false}
-TORAN_CRON_TIMER=${TORAN_CRON_TIMER:-fifteen}
-TORAN_CRON_TIMER_DAILY_TIME=${TORAN_CRON_TIMER_DAILY_TIME:-04:00}
-TORAN_TOKEN_GITHUB=${TORAN_TOKEN_GITHUB:-false}
-TORAN_TRACK_DOWNLOADS=${TORAN_TRACK_DOWNLOADS:-false}
-TORAN_MONO_REPO=${TORAN_MONO_REPO:-false}
-TORAN_SECRET=$(cat /dev/urandom | tr -dc 'a-zA-Z0-9' | fold -w 32 | head -n 1)
-
-# Checking Toran Proxy Configuration
-if [ "${TORAN_HTTPS}" = "true" ]; then
-    TORAN_SCHEME="https"
-elif [ "${TORAN_HTTPS}" = "false" ]; then
-    TORAN_SCHEME="http"
-else
-    echo "ERROR: "
-    echo "  Variable TORAN_HTTPS isn't valid ! (Values accepted : true/false)"
-    exit 1
-fi
-
-if [ "${TORAN_REVERSE}" != "true" ] && [ "${TORAN_REVERSE}" != "false" ]; then
-    echo "ERROR: "
-    echo "  Variable TORAN_REVERSE isn't valid ! (Values accepted : true/false)"
-    exit 1
-fi
-
-if [ "${TORAN_TRACK_DOWNLOADS}" != "true" ] && [ "${TORAN_TRACK_DOWNLOADS}" != "false" ]; then
-    echo "ERROR: "
-    echo "  Variable TORAN_TRACK_DOWNLOADS isn't valid ! (Values accepted : true/false)"
-    exit 1
-fi
-
-if [ "${TORAN_MONO_REPO}" != "true" ] && [ "${TORAN_MONO_REPO}" != "false" ]; then
-    echo "ERROR: "
-    echo "  Variable TORAN_MONO_REPO isn't valid ! (Values accepted : true/false)"
-    exit 1
-fi
-
-# Checking Cron time
-if [ "${TORAN_CRON_TIMER}" != "minutes" ] && [ "${TORAN_CRON_TIMER}" != "five" ] && [ "${TORAN_CRON_TIMER}" != "fifteen" ] && [ "${TORAN_CRON_TIMER}" != "half" ] && [ "${TORAN_CRON_TIMER}" != "hour" ] && [ "${TORAN_CRON_TIMER}" != "daily" ]; then
-    echo "ERROR: "
-    echo "  Variable TORAN_CRON_TIMER isn't valid ! (Values accepted : minutes/five/fifteen/half/hour)"
-    exit 1
-fi
-
-# Checking Cron daily time
-if [[ ! "${TORAN_CRON_TIMER_DAILY_TIME}" =~ ^[0-9]{2}:[0-9]{2}$ ]]; then
-    echo "ERROR: "
-    echo "  Variable TORAN_CRON_TIMER_DAILY_TIME isn't valid ! (Format accepted : HH:MM)"
-    exit 1
-fi
-
 # Load parameters toran
 cp -f $WORK_DIRECTORY/app/config/parameters.yml.dist $WORK_DIRECTORY/app/config/parameters.yml
 sed -i "s|toran_scheme:.*|toran_scheme: $TORAN_SCHEME|g" $WORK_DIRECTORY/app/config/parameters.yml